write test class
